Refactor webapp to MySQL
