+++ dockerlab/webapp/package.json
@@ -15,8 +15,9 @@
     "@fakerjs/animal": "^2.2.0",
     "@hapi/boom": "^10.0.1",
     "express": "^4.18.2",
-    "mongodb": "^5.6.0",
+    "knex": "^3.0.1",
     "morgan": "^1.10.0",
+    "mysql2": "^3.6.2",
     "sqlite": "^4.2.1",
     "sqlite3": "^5.1.6"
   },
