+++ dockerlab/webapp/src/persistence/index.js
@@ -1,5 +1,5 @@
-import * as mongodb from './mongodb.js';
+import * as mysql from './mysql.js';
 import * as sqlite from './sqlite.js';
 
-const database = process.env.MONGO_URL ? mongodb : sqlite;
+const database = process.env.MYSQL_URL ? mysql : sqlite;
 export default database;
